refactor: Rename memory_repository to platform_repository and implement MemoryRepository trait
